Renamed TextInputView and added some comments to my code.
I renamed the TextInputView to TextSelectView and the TextInput component to be TextSelect because in the future I want to add a TextInput component to behave like the NumberSelect component. (Which I plan on renaming to be NumberInput to fit the others. Because you aren't really selecting a number, you are providing one.